Move payload implementations into "payloads" sub-directory
