+++ src/index.ts
@@ -39,12 +39,12 @@ export * from './lib/crypto_wrappers/cms/envelopedData';
 export { SignatureOptions } from './lib/crypto_wrappers/cms/signedData';
 
 // RAMF
-export { default as Payload } from './lib/messages/PayloadPlaintext';
+export { default as Payload } from './lib/messages/payloads/PayloadPlaintext';
 export { default as RAMFError } from './lib/ramf/RAMFError';
 export { default as RAMFSyntaxError } from './lib/ramf/RAMFSyntaxError';
 export { default as Message } from './lib/messages/Message';
 export { default as Parcel } from './lib/messages/Parcel';
-export { default as ServiceMessage } from './lib/messages/ServiceMessage';
+export { default as ServiceMessage } from './lib/messages/payloads/ServiceMessage';
 export { default as Cargo } from './lib/messages/Cargo';
 export { CargoCollectionAuthorization } from './lib/messages/CargoCollectionAuthorization';
 export { default as CargoMessageSet } from './lib/messages/payloads/CargoMessageSet';
